filter functionality now works for all MCSP tables. Need to fix styling and get the ETS dates to sort.
